+++ CHANGELOG.md
@@ -5,11 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- connection issue
 
 ## [0.2.0] - 2019-09-24
 ### Added
 - database migrations
-
 
 ## [0.1.0] - 2019-09-18
 ### Added